refactor: simplify overlay
